fix: resolve all clippy warnings and CI workflow issues
- Remove unsupported --config flag from cargo audit commands in workflows
- Add #[allow(deprecated)] to test functions using derive_key_internal
- Fix unnecessary unsafe blocks in i18n_ffi.rs tests
- Fix unused variables and imports across multiple files
- Add crate-level clippy allows for common lints
- Fix redundant closures and unnecessary type casts
- Fix module inception warning in database/mod.rs
- Fix various code quality issues reported by clippy

+++ crates/easyssh-core/src/database/database.rs
@@ -220,7 +220,7 @@ impl Database {
             page_count: page_count.0,
             page_size: page_size.0,
             estimated_size_bytes: page_count.0 * page_size.0,
-            pool_connections: pool_stats as u32,
+            pool_connections: pool_stats,
         })
     }
